update admin & fix bug generate certif

+++ app/Http/Controllers/superadmin/EventController.php
@@ -47,6 +47,7 @@ class EventController extends Controller
      */
     public function store(Request $request)
     {
+        // Validate request
         $this->validate($request, [
             'nama_event' => 'required|string|max:255',
             'email' => 'required|email|max:255',
@@ -58,44 +59,52 @@ class EventController extends Controller
             'user_id' => 'required|string|max:255',
         ]);
     
+        
         $logoImage = $request->file('logo');
         if ($logoImage && $logoImage->getSize() > 2048 * 1024) {
             return redirect()->back()->with('error', 'Logo Harus Kurang Dari 2MB');
         }
-
+    
+        
         $ttdImage = $request->file('ttd');
         if ($ttdImage && $ttdImage->getSize() > 2048 * 1024) {
             return redirect()->back()->with('error', 'Tanda tangan Harus Kurang Dari 2MB');
         }
-
-
+    
+    
         $logoPath = null;
         if ($request->hasFile('logo')) {
             $logoImage = $request->file('logo');
             $logoImagePath = $logoImage->getPathname();
             
             $outputLogoPath = storage_path('app/public/logos/removed_' . $logoImage->getClientOriginalName());
-
-            $removeBg = new RemoveBg(config('removebg.api_key'));
-            $removeBg->file($logoImagePath)->save($outputLogoPath);
-            
-            $logoPath = '/logos/removed_' . $logoImage->getClientOriginalName();
+    
+            try {
+                $removeBg = new RemoveBg(config('removebg.api_key'));
+                $removeBg->file($logoImagePath)->save($outputLogoPath);
+                $logoPath = '/logos/removed_' . $logoImage->getClientOriginalName();
+            } catch (\Exception $e) {
+                \Log::error('RemoveBg error for logo' . $e->getMessage());
+                return redirect()->back()->with('error', 'Gagal remove background logo');
+            }
         }
-
         
         $ttdPath = null;
         if ($request->hasFile('ttd')) {
             $ttdImage = $request->file('ttd');
             $ttdImagePath = $ttdImage->getPathname();
-
+    
             $outputTtdPath = storage_path('app/public/ttd/removed_' . $ttdImage->getClientOriginalName());
-        
-            $removeBg = new RemoveBg(config('removebg.api_key')); 
-            $removeBg->file($ttdImagePath)->save($outputTtdPath);
-        
-            $ttdPath = '/ttd/removed_' . $ttdImage->getClientOriginalName();
+    
+            try {
+                $removeBg = new RemoveBg(config('removebg.api_key'));
+                $removeBg->file($ttdImagePath)->save($outputTtdPath);
+                $ttdPath = '/ttd/removed_' . $ttdImage->getClientOriginalName();
+            } catch (\Exception $e) {
+                \Log::error('RemoveBg error for ttd' . $e->getMessage());
+                return redirect()->back()->with('error', 'Gagal remove background tanda tangan:');
+            }
         }
-        
     
         
         Event::create([
@@ -105,12 +114,13 @@ class EventController extends Controller
             'deskripsi' => $request->deskripsi,
             'logo' => $logoPath,
             'tanggal' => $request->tanggal,
-            'ttd' => $ttdPath,          
+            'ttd' => $ttdPath,
             'user_id' => $request->user_id,
         ]);
     
-        return redirect()->route('superadmin.event');
+        return redirect()->route('superadmin.event')->with('success', 'Event created successfully!');
     }
+    
 
     /**
      * Display the specified resource.
@@ -154,54 +164,64 @@ class EventController extends Controller
             'email' => 'required|email|max:255',
             'no_telp' => 'required|numeric',
             'deskripsi' => 'nullable|string',
-            'logo' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048',
             'tanggal' => 'required|date',
             'ttd' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048',
+            'logo' => 'nullable|image|mimes:jpeg,png,jpg,gif|max:2048',
             'user_id' => 'required|string|max:255',
         ]);
-    
-        // Cari data event berdasarkan ID
+
+        $logoImage = $request->file('logo');
+        if ($logoImage && $logoImage->getSize() > 2048 * 1024) {
+            return redirect()->back()->with('error', 'Logo Harus Kurang Dari 2MB');
+        }
+
+        $ttdImage = $request->file('ttd');
+        if ($ttdImage && $ttdImage->getSize() > 2048 * 1024) {
+            return redirect()->back()->with('error', 'Tanda tangan Harus Kurang Dari 2MB');
+        }
+
         $event = Event::findOrFail($id);
-    
-        // Mengolah logo
+
         $logoPath = $event->logo;
         if ($request->hasFile('logo')) {
             if ($event->logo) {
-                // Hapus logo lama
                 Storage::disk('public')->delete($event->logo);
             }
-    
-            // Proses logo baru dengan menghapus background
+
             $logoImage = $request->file('logo');
             $logoImagePath = $logoImage->getPathname();
             $outputLogoPath = storage_path('app/public/logos/removed_' . $logoImage->getClientOriginalName());
-            
-            $removeBg = new RemoveBg(config('removebg.api_key'));
-            $removeBg->file($logoImagePath)->save($outputLogoPath);
-    
-            $logoPath = '/logos/removed_' . $logoImage->getClientOriginalName();
+
+            try {
+                $removeBg = new RemoveBg(config('removebg.api_key'));
+                $removeBg->file($logoImagePath)->save($outputLogoPath);
+                $logoPath = '/logos/removed_' . $logoImage->getClientOriginalName();
+            } catch (\Exception $e) {
+                \Log::error('RemoveBg error for logo ');
+                return redirect()->back()->with('error', 'Gagal remove background logo');
+            }
         }
-    
-        // Mengolah ttd
+
         $ttdPath = $event->ttd;
         if ($request->hasFile('ttd')) {
             if ($event->ttd) {
-                // Hapus ttd lama
                 Storage::disk('public')->delete($event->ttd);
             }
-    
-            // Proses ttd baru dengan menghapus background
+
             $ttdImage = $request->file('ttd');
             $ttdImagePath = $ttdImage->getPathname();
             $outputTtdPath = storage_path('app/public/ttd/removed_' . $ttdImage->getClientOriginalName());
-            
-            $removeBg = new RemoveBg(config('removebg.api_key'));
-            $removeBg->file($ttdImagePath)->save($outputTtdPath);
-    
-            $ttdPath = '/ttd/removed_' . $ttdImage->getClientOriginalName();
+
+            try {
+                $removeBg = new RemoveBg(config('removebg.api_key'));
+                $removeBg->file($ttdImagePath)->save($outputTtdPath);
+                $ttdPath = '/ttd/removed_' . $ttdImage->getClientOriginalName();
+            } catch (\Exception $e) {
+                \Log::error('RemoveBg error for ttd');
+                return redirect()->back()->with('error', 'Gagal remove background tanda tangan');
+            }
         }
-    
-        // Update data event
+
         $event->update([
             'nama_event' => $request->nama_event,
             'email' => $request->email,
@@ -212,10 +232,10 @@ class EventController extends Controller
             'ttd' => $ttdPath,
             'user_id' => $request->user_id,
         ]);
-    
-        // Redirect ke halaman event dengan pesan sukses
+
         return redirect()->route('superadmin.event')->with('success', 'Event berhasil diperbarui!');
     }
+
